+++ Js2IL.Tests/Node/Net/JavaScript/Net_Socket_SetEncoding_Utf8.js
@@ -3,30 +3,34 @@
 const net = require("node:net");
 
 const server = net.createServer(function (socket) {
-  let received = "";
+  socket.setEncoding("utf8");
 
+  let received = "";
   socket.on("data", function (chunk) {
-    received += chunk.toString("utf8");
+    console.log("server chunk:" + typeof chunk + ":" + chunk);
+    received += chunk;
   });
 
   socket.on("end", function () {
     console.log("server recv:" + received);
-    socket.write("pong:" + received.toUpperCase());
-    socket.end();
+    socket.end("done:" + received);
   });
 });
 
 server.listen(0, "127.0.0.1", function () {
   const info = server.address();
-  console.log("listening:" + info.family + ":" + info.address + ":" + (info.port > 0));
-
   const client = net.connect(info.port, info.address, function () {
-    client.end("ping");
+    client.write(Buffer.from([0xE2, 0x82]));
+    setTimeout(function () {
+      client.end(Buffer.from([0xAC, 0x21]));
+    }, 0);
   });
 
+  client.setEncoding("utf8");
   let reply = "";
   client.on("data", function (chunk) {
-    reply += chunk.toString("utf8");
+    console.log("client chunk:" + typeof chunk + ":" + chunk);
+    reply += chunk;
   });
 
   client.on("end", function () {